Merge foundation into master

+++ models/upload/storable.rb
@@ -1,12 +1,9 @@
 module Upload::Storable
 
-  def delete
-    if thumbnail_id
-      thumbnail = group.uploads.find(thumbnail_id)
-      thumbnail.destroy
-    end
+  def destroy
+    thumbnail.destroy if thumbnail
     dir = File.join('uploads', id.to_s)
-    FileUtils.rm_rf(dir)
+    `srm -r #{dir}`
     super
   end
